Use fully qualified name for test case and class name
In file view, Azure uses the class name to categorize test cases. In
the case of Caraya, the fully qualified VI name reflects the "class"
containing the assertions. As such, the fully qualified name should
be used for both the class name and test case.

In the case of the test case, the assertion label is used to further
distinguish between test cases. This, however, only works when the
user provides a custom label. Otherwise, the same label may occur
multiple times.