JUnit Conversion: Rebuild 'Get Error Details from Error Cluster.vi'
We previously assumed that all errors have the same fundamental structure
using <ERR> and <APPEND> tags to indicate specific error kinds. This is
not the case as many errors do not include either <ERR> or <APPEND> tags
and may have a different structure altogether.

This is a new approach at error conversion, which provides special cases
for exceptions from .NET (error 1172) and Call Library Function Node
(error 1097).

In the case of a .NET exception, it attempts to extract the exception
type and caller. These are assigned to 'type' and 'stack trace', where
as the exception message is assigned to 'message'. For example,

  Message: "Constructor Node Error creating instance of String in assembly
            System.String, mscorlib, Version=4.0.0.0, Culture=neutral,
            PublicKeyToken=b77a5c561934e089,
            (System.Reflection.AmbiguousMatchException: Ambiguous match
            found.)"
  Type: "System.Reflection.AmbiguousMatchException"
  Stack Trace: "Method.vi"

CLFN exceptions are lightweight and it attempts to extract the type
and stack trace. For the error message, the default CLFN exception
message is reported. For example,

  Message: "LabVIEW:  An exception occurred within the external code
            called by a Call Library Function Node. The exception might
            have corrupted the LabVIEW memory. Save any work to a new
            location and restart LabVIEW."
  Type: "Call Library Function Node"
  Stack Trace: "Method.vi"

Finally, for any other error code, the default converter is used.
It assumes that the error was generated with the standard function
'Error Cluster From Error Code'.

When the error source contains an <APPEND> tag, its value is assumed
to contain the stack trace. The header - which is typically included
in the error source - is removed.

When the <ERR> tag exists, it is assumed that the <ERR> tag either
contains only the error messsage or the error message and the stack
trace. The latter is detected from the header. The stack trace from
the <ERR> tag is only used when the <APPEND> tag has no stack info.

In the case that the <ERR> case is not present, it attempts to extract
the type info and if that fails, reports the entire string as error
message. In the case that no error message is found, it attempts to
lookup the error code in the error dictionary. That way, custom error
codes are taken into account. The type info is also extracted from the
source string before the <ERR> tag.

For example,

  Message: "My error message."
  Type: "Method.vi"
  Stack Trace: "Method.vi
                Caller.vi"

All test cases are updated and new test cases for .NET and CLFN exceptions
added. Some test cases that are incompatible with the new implementation
were removed.